games/diceladders/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
requests
loguru
Pillow

